Colour DT column by range (automatic range)

Colours DT columns based on the minimum and maximum of each column. Defaults to a purple-green colour scheme.

fmt_dtcol(
  dt,
  df,
  col_id,
  column_range = NULL,
  colours = c("#488F31", "#F1F1F1", "#783DBA")
)

Arguments

dt

a DT object created by DT::datatable()

df

the original dataframe used to create the DT

col_id

a vector of column identifiers (number or name)

column_range

Either the minimum and maximum range to be coloured, or "Auto" which defaults to the max and min of the table column.

colours

A set of hex colours to be passed on to colorRampPalette()

Value

A DT object with added colour formatting

Examples

DT::datatable(mtcars) |>
  fmt_dtcol(df = mtcars, column_range = NULL, col_id = c("mpg", "hp"))