Get User's Reported Time Zone

Uses some Javascript to pass the client's timezone to the Shiny session.

use_client_tz(inputId = "_client_tz")

get_client_tz(
  inputId = "_client_tz",
  session = shiny::getDefaultReactiveDomain()
)

Arguments

inputId

character string, name of shiny input. Defaults to "_client_tz". When used in use_client_tz, may need to be namespaced if called within a module.

session

a shiny::session object, defaults to shiny::getDefaultReactiveDomain

Value

use_client_tz(): HTML tags to be included in Shiny UI

timezone

Examples

if(interactive()){

shiny::shinyApp(
  ui = shiny::fluidPage(use_client_tz(), shiny::textOutput("time")),
  server = function(input, output, session) {
    # can be used outside of reactive context, if desired
    tz <- get_client_tz()
    time <- format(Sys.time(), format = "%x %X", tz = tz)
    output$time <- renderText(paste0(tz,": ", time))
  }
)
}