Generate Shiny Inputs for a vector of identifiers

Generate Shiny Inputs for a vector of identifiers

gen_input_map(uid, FUN, id_prefix = NULL, ...)

Arguments

uid

a vector of unique identifiers

FUN

a ShinyInput function - should theoretically work for all shinyInput functions that have inputid as the first argument

id_prefix

a string that will be combined with the unique identifier and passed as the inputid

...

other arguments to be passed to the FUN

Value

a character vector of shinyInputs

Examples

gen_input_map(1:5, shiny::numericInput, id_prefix = "playerid_", label = "my_label", value = 1)
#> [1] "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"playerid_1-label\" for=\"playerid_1\">my_label</label>\n  <input id=\"playerid_1\" type=\"number\" class=\"form-control\" value=\"1\" min=\"playerid_1\"/>\n</div>"
#> [2] "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"playerid_2-label\" for=\"playerid_2\">my_label</label>\n  <input id=\"playerid_2\" type=\"number\" class=\"form-control\" value=\"1\" min=\"playerid_2\"/>\n</div>"
#> [3] "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"playerid_3-label\" for=\"playerid_3\">my_label</label>\n  <input id=\"playerid_3\" type=\"number\" class=\"form-control\" value=\"1\" min=\"playerid_3\"/>\n</div>"
#> [4] "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"playerid_4-label\" for=\"playerid_4\">my_label</label>\n  <input id=\"playerid_4\" type=\"number\" class=\"form-control\" value=\"1\" min=\"playerid_4\"/>\n</div>"
#> [5] "<div class=\"form-group shiny-input-container\">\n  <label class=\"control-label\" id=\"playerid_5-label\" for=\"playerid_5\">my_label</label>\n  <input id=\"playerid_5\" type=\"number\" class=\"form-control\" value=\"1\" min=\"playerid_5\"/>\n</div>"
gen_input_map(FUN = numericInput, id_prefix = "itemprice_", uid = c(1, 3, 5, 7), ...)
#> Error in eval(expr, envir, enclos): '...' used in an incorrect context