Progressively

This function helps add progress-reporting to any function - given function f() and progressor p(), it will return a new function that calls f() and then (on-exiting) will call p() after every iteration. Now superseded by purrr's map .progress arguments.

progressively(f, p = NULL)

Arguments

f

a function to add progressr functionality to.

p

a progressor function similar to that created by progressr::progressor()

Value

a function that does the same as f but it calls p() after iteration.

Details

This is inspired by purrr's safely, quietly, and possibly function decorators.

See also

Examples

# \donttest{
read_rosters <- function(){
  urls <- c("https://github.com/nflverse/nflfastR-roster/raw/master/data/seasons/roster_2020.csv",
            "https://github.com/nflverse/nflfastR-roster/raw/master/data/seasons/roster_2021.csv")

  p <- progressr::progressor(along = urls)
  lapply(urls, progressively(read.csv, p))
}

progressr::with_progress(read_rosters())
#> Warning: cannot open URL 'https://github.com/nflverse/nflverse-rosters/raw/master/data/seasons/roster_2020.csv': HTTP status was '404 Not Found'
#> Error in file(file, "rt"): cannot open the connection to 'https://github.com/nflverse/nflfastR-roster/raw/master/data/seasons/roster_2020.csv'
# }