Unbind Shiny Inputs (UI function)

When generating inputs reactively, it's sometimes necessary to unbind old inputs so that you can use the updated inputs in Shiny. `unbind_dt_js() adds the JS to the UI, while unbind_dt() is called as needed in the server component (usually as part of an observeEvent or eventReactive)

unbind_dt_js()

Value

Adds JS to HTML head

Examples

unbind_dt_js()

#>